refactor(List): move exports
getDate & formatLocaleDateString -> date.js
getChecks & getLastUpdatedData -> App.js
